Factor out TimeWindowTrigger.
