+++ airbyte-cdk/bulk/core/base/src/main/resources/application.yaml
@@ -6,4 +6,5 @@ airbyte:
     disk:
       bytes: ${CONNECTOR_STORAGE_LIMIT_BYTES:5368709120} # 5GB
   flush:
-    interval-ms: 900000 # 15 minutes
+    rate-ms: 900000 # 15 minutes
+    window-ms: 900000 # 15 minutes
