+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/task/internal/FlushTickTask.kt
@@ -1,4 +1,4 @@
-package io.airbyte.cdk.load.task
+package io.airbyte.cdk.load.task.internal
 
 import io.airbyte.cdk.load.command.DestinationCatalog
 import io.airbyte.cdk.load.command.DestinationStream
@@ -7,23 +7,27 @@ import io.airbyte.cdk.load.message.DestinationRecordWrapped
 import io.airbyte.cdk.load.message.MessageQueueSupplier
 import io.airbyte.cdk.load.message.StreamFlushTickMessage
 import io.airbyte.cdk.load.state.Reserved
+import io.airbyte.cdk.load.task.KillableScope
+import io.airbyte.cdk.load.task.SyncLevel
 import io.micronaut.context.annotation.Value
 import jakarta.inject.Singleton
+import java.time.Clock
 
 @Singleton
 class FlushTickTask(
-    @Value("\${airbyte.flush.interval-ms}") private val tickIntervalMs: Long,
-    private val timeUtils: TimeProvider,
+    @Value("\${airbyte.flush.rate-ms}") private val tickIntervalMs: Long,
+    private val clock: Clock,
+    private val coroutineTimeUtils: TimeProvider,
     private val catalog: DestinationCatalog,
     private val recordQueueSupplier: MessageQueueSupplier<DestinationStream.Descriptor, Reserved<DestinationRecordWrapped>>,
-): SyncLevel {
+): SyncLevel, KillableScope {
     override suspend fun execute() {
         while (true) {
-            timeUtils.delay(tickIntervalMs)
+            coroutineTimeUtils.delay(tickIntervalMs)
 
             catalog.streams.forEach {
                 val queue = recordQueueSupplier.get(it.descriptor)
-                queue.publish(Reserved(value = StreamFlushTickMessage(timeUtils.currentTimeMillis())))
+                queue.publish(Reserved(value = StreamFlushTickMessage(clock.millis())))
             }
         }
     }
